requirements.txt
# Core Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Sources
yfinance>=0.2.0
alpaca-trade-api>=3.0.0
ccxt>=4.0.0

# Technical Analysis
ta>=0.10.2
# talib-binary>=0.4.0  # Using TA-Lib instead
pandas-ta>=0.3.14b0

# Machine Learning
scikit-learn>=1.3.0
ydf>=0.6.0
tensorflow-decision-forests>=1.5.0

# Backtesting & Performance
backtrader>=1.9.78.123
pyfolio>=0.9.2
quantstats>=0.0.62

# Development Tools
jupyter>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Risk Management
volatility>=2.4.1
arch>=5.6.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
